Adding task edition and prompting confirmation before delete

+++ app/src/views/TaskList.vue
@@ -42,14 +42,11 @@
 
                     <v-list-item-content>
                       <v-list-item-title :class="{'grey--text text--lighten-1': task.is_closed}">{{ task.description }}</v-list-item-title>
-                      <v-list-item-subtitle :class="{'grey--text text--lighten-1': task.is_closed}">Created {{ task.created_at | str2date | relativeTime }} <v-chip x-small>{{ task.project.name }}</v-chip></v-list-item-subtitle>
+                      <v-list-item-subtitle :class="{'grey--text text--lighten-1': task.is_closed}">Created {{ task.created_at | str2date | relativeTime }} <v-chip x-small>{{ projectKeyVal[task.project_id] }}</v-chip></v-list-item-subtitle>
                     </v-list-item-content>
 
                     <v-list-item-avatar>
-                      <v-btn fab color="red" :loading="task.deleting" @click="removeTask(task, i)" v-if="!task.is_closed">
-                        <v-icon class="white--text">mdi-delete</v-icon>
-                      </v-btn>
-                      <v-tooltip left v-else>
+                      <v-tooltip left v-if="task.is_closed">
                         <template v-slot:activator="{ on }">
                           <v-avatar color="green" v-on="on">
                             <v-icon class="white--text">mdi-check</v-icon>
@@ -57,6 +54,31 @@
                         </template>
                         <span>Closed at {{ task.finished_at | str2date | formatdate('MMMM Do YYYY, h:mm:ss a') }}</span>
                       </v-tooltip>
+
+                      <v-menu v-else bottom left>
+                        <template v-slot:activator="{ on }">
+                          <v-btn icon v-on="on">
+                            <v-icon>mdi-dots-vertical</v-icon>
+                          </v-btn>
+                        </template>
+
+                        <v-list>
+                          <v-list-item>
+                            <v-list-item-title>
+                              <v-btn text small :loading="task.deleting" @click="removeTask(task, i)">
+                                <v-icon class="red--text">mdi-delete</v-icon> Delete
+                              </v-btn>
+                            </v-list-item-title>
+                          </v-list-item>
+                          <v-list-item>
+                            <v-list-item-title>
+                              <v-btn  text small :loading="task.editing" @click="editTask(task)">
+                                <v-icon class="primary--text">mdi-pen</v-icon> Edit
+                              </v-btn>
+                            </v-list-item-title>
+                          </v-list-item>
+                        </v-list>
+                      </v-menu>
 
                     </v-list-item-avatar>
                 </v-list-item>
@@ -69,12 +91,16 @@
         </v-card>
       </v-col>
     </v-row>
+    <v-confirm ref="confirm"></v-confirm>
+    <task-edit ref="taskedit"></task-edit>
   </v-container>
 </template>
 
 <script>
 
   import { mapGetters } from 'vuex'
+  import Confirm from '../components/util/Confirm'
+  import TaskEdit from '../components/app/TaskEdit'
 
   export default {
 
@@ -88,12 +114,24 @@
       }
     }),
 
+    components: {
+      'v-confirm': Confirm,
+      'task-edit': TaskEdit
+    },
+
     computed: {
       ...mapGetters([
         'taskList',
         'projectList',
         'isAuthenticated'
       ]),
+      projectKeyVal() {
+        let map = {}
+        this.$store.getters.projectList.forEach(project => {
+          map[project.id] = project.name
+        })
+        return map
+      }
     },
 
     methods: {
@@ -126,17 +164,25 @@
       },
 
       removeTask(task, index) {
-        this.$set(task,'deleting', true)
-        this.$http({ url: `/task/${task.id}`, method: 'delete'}).then(response => {
-          this.$showMessage(response.data.message, 'success')
-          this.taskList.splice(index, 1);
-          this.$delete(task,'deleting')
-        }).catch(error => {
-          this.$delete(task,'deleting')
-          let data = error.response.data || {}
-          this.$showMessage(data.message || 'Fail to close task', 'error')
+        this.$refs.confirm.open('Are you sure to delete?', 'This record can\'t be recovered', { color: 'error' }).then(confirm => {
+          if(confirm) {
+            this.$set(task,'deleting', true)
+            this.$http({ url: `/task/${task.id}`, method: 'delete'}).then(response => {
+              this.$showMessage(response.data.message, 'success')
+              this.taskList.splice(index, 1);
+              this.$delete(task,'deleting')
+            }).catch(error => {
+              this.$delete(task,'deleting')
+              let data = error.response.data || {}
+              this.$showMessage(data.message || 'Fail to close task', 'error')
+            })
+          }
         })
       },
+
+      editTask(task) {
+        this.$refs.taskedit.open(task)
+      }
 
     },
   }
